Move auth token logic out of web player module
