+++ app/static/player.mjs
@@ -1,57 +1,24 @@
-/*
-    This script is based on Spotify developers tutorials:
-    - https://developer.spotify.com/documentation/web-playback-sdk/quick-start/#
-    - https://developer.spotify.com/documentation/web-playback-sdk/reference/#playing-a-spotify-uri
-
-    Sets up Spotify web player and exposes play method to use the player.
-*/
-
-const getBearerTokenFromUrl = () => {
-    // Adapted from https://glitch.com/edit/#!/spotify-implicit-grant
-    // get bearer token
-    const hash = window.location.hash
-        .substring(1)
-        .split('&')
-        .reduce(function (initial, item) {
-            if (item) {
-                var parts = item.split('=');
-                initial[parts[0]] = decodeURIComponent(parts[1]);
-            }
-            return initial;
-    }, {});
-    window.location.hash = '';  // clean up URL
-    return !hash.access_token? undefined : hash.access_token;
-}
-let bearerToken = getBearerTokenFromUrl();
-
-
-const redirectToLogin = ({authEndpointTemplate, clientId, authScopes, redirectUrl}) => {
-    // Redirect to Spotify authorization
-    window.location = encodeURI(authEndpointTemplate.For({
-        clientId: clientId,
-        scopes: authScopes,
-        redirectUrl: redirectUrl
-    }));
-}
-
 let mostRecentTrackUri = undefined, player = undefined;
 const Player = {
     IsConnected: false,
-    OnReady: () => { console.log("it's ready!!"); },
-    _OnReady: ({ device_id }) => {
+    OnReady: ({ device_id }) => {
         console.log('Ready with Device ID', device_id);
-        Player.OnReady();
     },
     OnSongChange: ({songName, artistName, albumName, albumArtLink, songLink}) => {
         console.log("Now Playing " + songName + " by " + artistName);
-    }
+    },
+    BearerToken: undefined
 }
 
 Player.Init = () => {
-    // Spotify is linked in HTML page
+    if (Player.BearerToken === undefined) {
+        console.log("ERROR: Cannot initialize Player without bearer token.")
+        return;
+    }
+
     player = new Spotify.Player({
-      name: 'Muze',
-      getOAuthToken: cb => { cb(bearerToken); }
+      name: 'Muze Radio',
+      getOAuthToken: cb => { cb(Player.BearerToken); }
     });
 
     player.addListener('initialization_error', ({ message }) => {
@@ -67,7 +34,7 @@ Player.Init = () => {
         console.error(`Received a playback error from Spotify player: ${message}`);
     });
 
-    player.addListener('ready', Player._OnReady);
+    player.addListener('ready', Player.OnReady);
 
     player.addListener('not_ready', ({ device_id }) => {
       console.log('Device ID has gone offline', device_id);
@@ -84,6 +51,16 @@ Player.Init = () => {
                 albumArtLink: current_track['album']['images'][0]['url'],
                 songLink: `https://open.spotify.com/track/${current_track['id']}`
             })
+        }
+    });
+
+    player.connect().then((success) => {
+        if (success) {
+            console.log("Connected player!");
+            Player.IsConnected = true;
+        } else {
+            console.log("ERROR: could not connect player");
+            Player.IsConnected = false;
         }
     });
 };
@@ -128,38 +105,6 @@ Player.PlaySong = ({spotify_uri}) => {
                 'Authorization': `Bearer ${access_token}`
             },
         });
-    });
-}
-
-Player.Connect = ({authEndpointTemplate, clientId, authScopes, redirectUrl, onReady}) => {
-    if (Player.IsConnected === true) {
-        console.log("Player already connected.");
-        onReady();
-    }
-    if (player === undefined) {
-        console.log("ERROR: Cannot get current song because player is not initialized.");
-        return;
-    }
-    if (bearerToken === undefined) {
-        console.log("Must log in before the Spotify Player can connect. Redirecting..")
-        redirectToLogin({
-            authEndpointTemplate: authEndpointTemplate,
-            clientId: clientId,
-            authScopes: authScopes,
-            redirectUrl: redirectUrl
-        });
-        return;
-    }
-
-    Player.OnReady = onReady;
-    player.connect().then((success) => {
-        if (success) {
-            console.log("Connected player!");
-            Player.IsConnected = true;
-        } else {
-            console.log("ERROR: could not connect player");
-            Player.IsConnected = false;
-        }
     });
 }
